extremely minor changes

+++ fileBuffer.c
@@ -26,6 +26,8 @@
 **/
 #include <errno.h>
 #include <string.h>
+#include <stdint.h>
+#include <inttypes.h>
 
 #include "fileBuffer.h"
 #include "stringBuffer.h"
@@ -212,8 +214,8 @@ void mcc_ShiftLineLeftAndShrink(mcc_LogicalLine_t *line,
 #if MCC_DEBUG
 void printFileBuffer(mcc_FileBuffer_t *buffer)
 {
-   printf("----- FileBuffer ----- \nfilename:\t%s\nline_no:\t%u\nbufferIndex:\t%u\nchars_read:\t%lu\n",
-          buffer->filename, buffer->line_no, buffer->bufferIndex, buffer->chars_read);
+   printf("----- FileBuffer ----- \nfilename:\t%s\nline_no:\t%u\nbufferIndex:\t%u\nchars_read:\t%" PRIuPTR "\n",
+          buffer->filename, buffer->line_no, buffer->bufferIndex, (uintptr_t)buffer->chars_read);
    printf("Current Char:\t%d\n", buffer->buffer[buffer->bufferIndex]);
 }
 #endif
